+++ gmmzoo/core.py
