+++ gmmzoo/utils.py
